refactor: code order
