+++ package.json
@@ -3,7 +3,8 @@
   "version": "4.1.1",
   "description": "Module to load environment variables from a `.envuse` file. Ideal to load configurations from the environment system, transform values and configure default values.",
   "publishConfig": {
-    "access": "public"
+    "access": "public",
+    "tag": "beta"
   },
   "engines": {
     "node": ">=14"
