+++ tsconfig.esm.json
@@ -5,7 +5,8 @@
     "module": "ES2020",
     "declaration": true,
     "sourceMap": true,
+    "rootDir": "src",
     "outDir": "esm"
   },
-  "files": ["utils/dtsfile.mts", "envuse.mts"]
+  "include": ["src/**/*"]
 }
